% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{abs_cutoff_adapt_COSMIC_df}
\alias{abs_cutoff_adapt_COSMIC_df}
\title{Optimal adapted cutoffs for absolute exposures of COSMIC signatures}
\usage{
data(abs_cutoff_adapt_COSMIC)
}
\description{
Data frame of signature specific cutoff values for the LCD analysis with
absolute exposures. All values represent optimal cutoffs. The optimal
cutoffs were determined for different choices of parameters in the cost
function of the optimization. The row index is equivalent to the ratio
between costs for false negative attribution and false positive
attribution. The columns correspond to the different signatures. To be
used with \code{\link{LCD_complex_cutoff}}. Based on the signatures available
at http://cancer.sanger.ac.uk/cosmic/signatures .
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{cutoffCosmicValid_abs_df}}

\code{\link{cutoffInitialValid_rel_df}}
}

