% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{add_as_fist_to_list}
\alias{add_as_fist_to_list}
\title{Add an element as first entry to a list}
\usage{
add_as_fist_to_list(in_list, in_element)
}
\arguments{
\item{in_list}{List to which an element is to be added}

\item{in_element}{Element to be added}
}
\value{
List with input element as first entry.
}
\description{
Works for all types of lists and inputs
}
\examples{
NULL
}

