% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{adjust_number_of_columns_in_list_of_catalogues}
\alias{adjust_number_of_columns_in_list_of_catalogues}
\title{Extend a list of mutational catalogues to consistent dimensions}
\usage{
adjust_number_of_columns_in_list_of_catalogues(in_all_list,
  in_merged_results_list)
}
\arguments{
\item{in_all_list}{A list of \code{s} lists, where \code{s is the number of strata}.
Every entry corresponds to one result of
\code{\link{create_mutation_catalogue_from_df}}}

\item{in_merged_results_list}{Data structure from an analysis without stratification, used here only
to get the correct dimensions, especially the correct number of columns.}
}
\value{
A list of numerical data frames with the adjusted stratified
mutational catalogues
}
\description{
In the stratification procedure, after having run
\code{\link{stratify_and_create_mutational_catalogue}}, the dimensions,
especially the number of columns (corresponding to the number of PIDs
or samples) of the catalogues for the different strata need not agree.
This function accounts for that and adds zero columns where necessary.
}
\seealso{
\code{\link{stratify_and_create_mutational_catalogue}}
}

