% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{aggregate_exposures_by_category}
\alias{aggregate_exposures_by_category}
\title{Aggregate exposures by category}
\usage{
aggregate_exposures_by_category(in_exposures_df, in_sig_ind_df, in_category)
}
\arguments{
\item{in_exposures_df}{Input data frame of exposures.}

\item{in_sig_ind_df}{Input data frame of meta information on the signatures. Has to match the
signatures in \code{in_exposures_df}}

\item{in_category}{Category to be aggregated over}
}
\value{
A dataframe of aggregated exposures.
}
\description{
If a valid category (i.e. it matches to a category specified in
in_sig_ind_df) is supplied, then the exposures are aggregated over this
category.
}
\examples{
NULL
}
\seealso{
\code{\link{LCD_complex_cutoff}}
}

