% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{AlexCosmicArtif_sig_df}
\alias{AlexCosmicArtif_sig_df}
\title{Data for Cosmic sigs, including artifacts}
\usage{
data(sigs)
}
\value{
A data frame
}
\description{
Data frame of the updated signatures list maintained by Ludmil Alexandrov at
\url{http://cancer.sanger.ac.uk/cosmic/signatures}. The data was downloaded from
\url{http://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt}
and reformatted. The column names are \emph{AC} pasted to the number of the
signature, e.g. \emph{AC5}. The naming convention for the rows is as described
for \code{\link{AlexInitialArtif_sig_df}}.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
Alexandrov et al. (Nature 2013)
}
\seealso{
\code{\link{AlexInitialArtif_sig_df}}
}

