% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{AlexCosmicValid_sigInd_df}
\alias{AlexCosmicValid_sigInd_df}
\title{Meta-information on updated mutational signatures}
\usage{
data(sigs)
}
\description{
A data frame with as many rows as there are signatures in \code{AlexCosmicValid_sig_df}
and four columns: one for the signature name, one carrying an index, one
encoding the colour for visualization and one describing the asserted biological
process.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{AlexCosmicValid_sig_df}}
}

