% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{AlexInitialArtif_sig_df}
\alias{AlexInitialArtif_sig_df}
\title{Information on initially published mutational signatures}
\usage{
data(sigs)
}
\description{
Data frame of the signatures published initially by Alexandrov et al. (Nature 
2013). There are 27 signatures which constitute the columns, 22 of which were
validated by an orhtogonal sequencing technology. These 22 are in the first 22
columns of the data frame. The column names are \emph{A} pasted to the number
of the signature, e.g. \emph{A5}. The nonvalidated signatures have an
additional letter in their naming convention: either \emph{AR1} - \emph{AR3}
or \emph{AU1} - \emph{AU2}. The rownames are the features, i.e. an encoding of
the nucleotide exchanges in their trinucleotide context, e.g. \emph{C>A ACA}.
In total there are 96 different features and therefore 96 rows when dealing
with a trinucleotide context. This data was downloaded from
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/signatures.txt} and
reformatted.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
Alexandrov et al. (Nature 2013)
}
\seealso{
\code{\link{AlexInitialArtif_sigInd_df}}

\code{\link{AlexCosmicValid_sig_df}}

\code{\link{AlexInitialValid_sig_df}}
}

