% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{AlexInitialValid_sig_df}
\alias{AlexInitialValid_sig_df}
\title{Data for initial sigs, only validated}
\usage{
data(sigs)
}
\value{
A data frame
}
\description{
Data frame of only the validated signatures published initially by Alexandrov
et al. (Nature 2013), corresponding to the first 22 columns of
\code{AlexInitialArtif_sig_df}
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
Alexandrov et al. (Nature 2013)
}
\seealso{
\code{\link{AlexInitialArtif_sig_df}}
}

