% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{AlexInitialValid_sigInd_df}
\alias{AlexInitialValid_sigInd_df}
\title{Meta-info for initial sigs, only validated}
\usage{
data(sigs)
}
\description{
A data frame with as many rows as there are signatures in
\code{AlexInitialValid_sig_df} and several columns: 
\itemize{
 \item \code{sig}: signature name
 \item \code{index}: corresponding to the row index of the signature
 \item \code{colour}: colour for visualization in stacked barplots
 \item \code{process}: asserted biological process
 \item \code{cat.coarse}: categorization of the signatures according
  to the asserted biological processes at low level of detail
 \item \code{cat.medium}: categorization of the signatures according
  to the asserted biological processes at intermediate level of detail
 \item \code{cat.high}: categorization of the signatures according
  to the asserted biological processes at high level of detail
 \item \code{cat.putative}: categorization of the signatures according
  to the asserted biological processes based on clustering and inference
 }
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{AlexInitialArtif_sig_df}}

\code{\link{AlexInitialValid_sig_df}}
}

