% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{average_over_present}
\alias{average_over_present}
\title{Average a data frame over a specified dimension}
\usage{
average_over_present(in_df, in_dimension)
}
\arguments{
\item{in_df}{Data frame to be normalized}

\item{in_dimension}{Dimension along which the averaging will be carried out}
}
\value{
A vector of the means
}
\description{
If averaging over columns, zero rows (i.e. those with \code{rowSums=0})
are left out, if averaging over rows, zero columns (i.e. those with
\code{colSums=0}) are left out.
}
\examples{
test_df <- data.frame(matrix(c(1,2,3,0,5,2,3,4,0,6,0,0,0,0,0,4,5,6,0,7),ncol=4))
## 1. Average over non-zero rows:
average_over_present(test_df,1)
## 2. Average over non-zero columns:
average_over_present(test_df,2)
}

