% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{classify_indels}
\alias{classify_indels}
\title{INDEL function V1 - not compartible with AlexandrovSignatures}
\usage{
classify_indels(
  in_df,
  in_ALT.field = "ALT",
  in_REF.field = "REF",
  in_breaks = c(-Inf, -10, -3, 0, 2, 9, Inf),
  in_labels = c("del3", "del2", "del1", "in1", "in2", "in3")
)
}
\arguments{
\item{in_df}{Input data frame containing the variances in a vcf-like format}

\item{in_ALT.field}{Column number for alternitve field}

\item{in_REF.field}{Coloumn number for reference field}

\item{in_breaks}{Handed over to function cut}

\item{in_labels}{Handed over to function cut}
}
\value{
classVector, a factor vector of indel sizes
}
\description{
INDEL function V1 - not compartible with AlexandrovSignatures
}
\examples{
 NULL

}
