% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{compute_comparison_stat_df}
\alias{compute_comparison_stat_df}
\title{Extract statistical measures for entity comparison}
\usage{
compute_comparison_stat_df(in_sim_df)
}
\arguments{
\item{in_sim_df}{A similarity data frame as extracted by \code{\link{compare_to_catalogues}}}
}
\value{
A dataframe containing statistical measures, prepared for bar plot
}
\description{
Compare one mutational catalogue (e.g. of one index patient) to a list of
reference mutational catalogues (e.g. from the initial Alexandrov
puplication) by cosine similarities
}
\examples{
NULL
}

