% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{computeLogLik}
\alias{computeLogLik}
\title{Compute the loglikelihood}
\usage{
computeLogLik(in_vector, in_pdf = NULL, verbose = FALSE)
}
\arguments{
\item{in_vector}{Numeric vector of input values of which the loglikelihood is
computed.}

\item{in_pdf}{Probability distribution function, if NULL a normal
distribution is used.}

\item{verbose}{Verbose if \code{in_verbose=1}}
}
\value{
A numeric value (sum of the logarithms of the likelihoods of the
  input vector)
}
\description{
Compute the loglikelihood
}
\examples{
NULL

}
