% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{correct_rounded}
\alias{correct_rounded}
\title{Readjust the vector to it's original norm after rounding}
\usage{
correct_rounded(x, in_interval = c(0, 1))
}
\arguments{
\item{x}{vector to be rounded}

\item{in_interval}{Interval}
}
\value{
The adapted form of the input vector \code{x}.
}
\description{
After use of the function \code{\link{round_precision}} the norm of the 
input vector may have been altered by the rounding procedure. This 
function restores the norm by altering only the largest entry in the 
rounded vector (in order to create the least possible relative error).
}
\examples{
 NULL

}
