% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{COSMIC_subgroups_df}
\alias{COSMIC_subgroups_df}
\title{Subgroup information for test data for plot functions}
\usage{
data(lymphoma_cohort_LCD_results)
}
\description{
Subgroup information for the data stored in
\code{\link{lymphoma_Nature2013_COSMIC_cutoff_exposures_df}} and
\code{\link{rel_lymphoma_Nature2013_COSMIC_cutoff_exposures_df}}.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{lymphoma_Nature2013_COSMIC_cutoff_exposures_df}}

\code{\link{rel_lymphoma_Nature2013_COSMIC_cutoff_exposures_df}}
}

