% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{cutoffCosmicArtif_rel_df}
\alias{cutoffCosmicArtif_rel_df}
\title{Opt. cutoffs, rel exposures for the COSMIC sigs, including artifacts}
\usage{
data(cutoffs)
}
\value{
A data frame
}
\description{
Data frame of signature specific cutoff values for the LCD analysis with
relative exposures. All values represent optimal cutoffs. The optimal
cutoffs were determined for different choices of parameters in the cost
function of the optimization. The row index is equivalent to the ratio
between costs for false negative attribution and false positive
attribution. The columns correspond to the different signatures. To be
used with \code{\link{LCD_complex_cutoff}}. Based on the signatures available
at http://cancer.sanger.ac.uk/cosmic/signatures .
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{cutoffCosmicValid_abs_df}}

\code{\link{cutoffInitialValid_rel_df}}
}

