% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{cutoffInitialArtif_abs_df}
\alias{cutoffInitialArtif_abs_df}
\title{Opt. cutoffs, abs exposures for the initial sigs, including artifacts}
\usage{
data(cutoffs)
}
\description{
Data frame of signature specific cutoff values for the LCD analysis with
absolute exposures. All values represent optimal cutoffs. The optimal
cutoffs were determined for different choices of parameters in the cost
function of the optimization. The row index is equivalent to the ratio
between costs for false negative attribution and false positive
attribution. The columns correspond to the different signatures. To be
used with \code{\link{LCD_complex_cutoff}}. Based on the signatures published
in Alexandrov et al. (Nature 2013).
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\seealso{
\code{\link{cutoffCosmicValid_abs_df}}

\code{\link{cutoffInitialValid_rel_df}}
}

