% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{deriveSigInd_df}
\alias{deriveSigInd_df}
\title{Derive a signature_indices_df object}
\usage{
deriveSigInd_df(querySigs, subjectSigs, querySigInd = NULL, in_sort = FALSE)
}
\arguments{
\item{querySigs}{The signatures to compare to (given signatures).}

\item{subjectSigs}{The signatures to be compared (new signatures).
Alternatively this may be a complex object of type list and contain data
from different deconvolutions, each of which having a set of signaturen to
be compared.}

\item{querySigInd}{The object of type signature_indices_df (additional
informatio for a set of signatures) belonging to the set of known
signatures.}

\item{in_sort}{Whether to sort or not}
}
\value{
An object of type signature_indices_df (additional informatio for a
  set of signatures) belonging to the set of new signatures.
}
\description{
Derive a data frame of type signature_indices_df (additional information for
a set of signatures) from a set of given signatures for a set of new
signatures.
}
\examples{
 NULL

}
\seealso{
\code{\link{relateSigs}}
}
