% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{disambiguateVector}
\alias{disambiguateVector}
\title{Disambiguate a vector}
\usage{
disambiguateVector(in_vector)
}
\arguments{
\item{in_vector}{Input vector}
}
\value{
The disambiguated vector.
}
\description{
Add numbered suffixes to redundant entries in a vector
}
\examples{
 NULL

}
