% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{enhanced_barplot}
\alias{enhanced_barplot}
\title{Plot the exposures of a cohort as a ComplexHeatmap}
\usage{
enhanced_barplot(mat, col, anno, anno_color, ylab = NULL, title = "",
  in_labels = TRUE, in_barplot_borders = TRUE,
  in_column_anno_borders = FALSE)
}
\arguments{
\item{mat}{Numerical data frame encoding the exposures \code{H}, i.e. which
signature contributes how much to which PID (patient identifier or sample).}

\item{col}{A data frame containing meta information about the signatures}

\item{anno}{A data frame indicating which PID (patient or sample identifyier) belongs
to which subgroup}

\item{anno_color}{Index vector influencing the order in which the PIDs are going to be
displayed}

\item{ylab}{String indicating the column name in \code{in_subgroups_df} to take the
subgroup information from.}

\item{title}{Title for the plot to be created.}

\item{in_labels}{Whether or not to show the names of the samples.}

\item{in_barplot_borders}{Whether or not to show border lines in barplot}

\item{in_column_anno_borders}{Whether or not to draw separating lines between the fields in the annotation}
}
\value{
The function doesn't return any value.
}
\description{
The exposures \code{H}, determined by NMF or by \code{\link{LCD}}, are
displayed as a stacked barplot by calling \code{\link[ComplexHeatmap]{Heatmap}}.
The x-axis displays the PIDs (patient identifier or sample), the y-axis
the counts attributed to the different signatures with their respective
colours per PID. It is analogous to \code{\link{plot_exposures}}.
}
\details{
It might be necessary to install the newest version of the development branch
 of the packages \pkg{circlize} and \pkg{ComplexHeatmap} by Zuguang Gu:
 \code{devtools::install_github("jokergoo/circlize")}
 \code{devtools::install_github("jokergoo/ComplexHeatmap")}
}
\examples{
data(lymphoma_cohort_LCD_results)
 temp_anno_color <- aggregate(col~subgroup,data=COSMIC_subgroups_df,function(l) return(l[1]))
 my_anno_color <- temp_anno_color$col
 names(my_anno_color) <- temp_anno_color$subgroup
 enhanced_barplot(lymphoma_Nature2013_COSMIC_cutoff_exposures_df[,order(COSMIC_subgroups_df$index)],
                  chosen_signatures_indices_df$colour,
                  COSMIC_subgroups_df$subgroup[order(COSMIC_subgroups_df$index)],
                  my_anno_color)
}
\seealso{
\code{\link[ComplexHeatmap]{HeatmapAnnotation}}

\code{\link[ComplexHeatmap]{Heatmap}}

\code{\link[ComplexHeatmap]{decorate_heatmap_body}}

\code{\link{plot_exposures}}
}

