% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{enrichSigs}
\alias{enrichSigs}
\title{Compare to background distribution}
\usage{
enrichSigs(in_cohort_exposures_df, in_background_exposures_df, in_sig_df)
}
\arguments{
\item{in_cohort_exposures_df}{Numerical data frame of the exposures of the
cohort of interest.}

\item{in_background_exposures_df}{Numerical data frame of the exposures of
the background.}

\item{in_sig_df}{Numerical data frame encoding the mutational signatures.}
}
\value{
A data frame with counts and p-values from Fisher tests.
}
\description{
Compare exposures from an analysis of mutational signatures in a cohort of
interest to exposures computed in a background (e.g. the set of WES and WGS
samples from Alexandrov 2013).
}
\examples{
NULL

}
