% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleINDEL_YAPSA}
\alias{exampleINDEL_YAPSA}
\title{Data structures used in examples, Indel tests and the Indel signature
vignette of the YAPSA package.}
\description{
Data structures used in examples, Indel tests and the Indel signature
vignette of the YAPSA package.
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23945592}
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
