% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{exchange_colour_vector}
\alias{exchange_colour_vector}
\title{Colours codes for displaying SNVs}
\usage{
data(exchange_colour_vector)
}
\description{
Vector attributing colours to nucleotide exchanges used when displaying SNV
information, e.g. in a rainfall plot.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}

