% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exome_mutCatRaw_df}
\alias{exome_mutCatRaw_df}
\title{Example mutational catalog for the exome vignette}
\usage{
data(smallCellLungCancerMutCat_NatureGenetics2012)
}
\value{
A data fame in the layout of a SNV mutational catalog
}
\description{
\code{exome_mutCatRaw_df}: A data frame in the format of a SNV mutation 
catalog. The mutational catalog contains SNV variants from a cohort of 
small-cell lung cancer published by Rudin et al. (Nature Genetics 2012) 
which was later used in the de novo discovery analysis of mutational 
signatures in human cancer by Alexandrov et al. (Nature 2013).
}
\examples{
data(smallCellLungCancerMutCat_NatureGenetics2012)
head(exome_mutCatRaw_df)
dim(exome_mutCatRaw_df)

}
\references{
\url{https://www.nature.com/articles/ng.2405}
}
