% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_foreign.R
\name{extract_names_from_gene_list}
\alias{extract_names_from_gene_list}
\title{Return gene names from gene lists}
\usage{
extract_names_from_gene_list(in_KEGG_gene_list, l)
}
\arguments{
\item{in_KEGG_gene_list}{Gene list to extract names from}

\item{l}{Index of the gene to be extracted}
}
\value{
The gene name.
}
\description{
Return gene names from gene lists
}
\examples{
NULL
}
\seealso{
\code{\link[KEGGREST]{keggGet}}

\code{\link{build_gene_list_for_pathway}}
}

