% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{find_affected_PIDs}
\alias{find_affected_PIDs}
\title{Find samples affected}
\usage{
find_affected_PIDs(in_gene_list, in_gene_vector, in_PID_vector)
}
\arguments{
\item{in_gene_list}{List of genes in the pathway of interest.}

\item{in_gene_vector}{Character vector for genes annotated to SNVs as in \code{vcf_like_df}.}

\item{in_PID_vector}{Character vector for sample names annotated to SNVs as in
 \code{vcf_like_df}.}
}
\value{
A character vector of the names of the affected samples
}
\description{
Find samples affected by SNVs in a certain pathway
}
\examples{
NULL
}

