% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GenomeOfNl_raw}
\alias{GenomeOfNl_raw}
\title{Example data for the Indel vignette}
\usage{
data(GenomeOfNl_raw)
}
\value{
A data frame in a vcf-like format
}
\description{
\code{GenomeOfNl_raw}: A data frame contains the gemiline varinats of 
the dutch population.  carrying point mutation 
calls. It represents a subset of the data stored in
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/
somatic_mutation_data/Lymphoma B-cell/
Lymphoma B-cell_clean_somatic_mutations_for_signature_analysis.txt}.
In the file available under that link somatic point mutation calls from 
several samples are listed in a vcf-like format. One column encodes the 
sample the variant was found in.
}
\examples{
data(GenomeOfNl_raw)
head(GenomeOfNl_raw)
dim(GenomeOfNl_raw)

}
\references{
release version 5 \url{http://www.nlgenome.nl/?page_id=9}
}
