% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{heatmap_exposures}
\alias{heatmap_exposures}
\title{Create heatmap to cluster the PIDs according to their signature exposures}
\usage{
heatmap_exposures(in_exposures_df, in_subgroups_df, in_signatures_ind_df,
  in_method = "manhattan", in_subgroup_column = "subgroup",
  in_subgroup_colour_column = NULL, in_palette = NULL, in_cutoff = 0,
  in_filename = NULL, in_title = "")
}
\arguments{
\item{in_exposures_df}{Numerical data frame encoding the exposures \code{H}, i.e. which
signature contributes how much to which PID (patient identifier or sample).}

\item{in_subgroups_df}{A data frame indicating which PID (patient or sample identifyier) belongs
to which subgroup}

\item{in_signatures_ind_df}{A data frame containing meta information about the signatures, especially the
asserted colour}

\item{in_method}{Method of the clustering to be supplied to \code{\link{dist}}. Can be either
of: \code{euclidean}, \code{maximum}, \code{manhattan}, \code{canberra},
\code{binary} or \code{minkowski}}

\item{in_subgroup_column}{Indicates the name of the column in which the subgroup information
is encoded in \code{in_subgroups_df}}

\item{in_subgroup_colour_column}{Indicates the name of the column in which the colour information for
subgroups is encoded in \code{in_subgroups_df}. If NULL, a rainbow palette
is used instead.}

\item{in_palette}{Palette with colours or colour codes for the labels (the text) of the leaves
in the dendrogram. Typically one colour per subgroup. If none is specified, a
rainbow palette of the length of the number of subgroups will be used as
default.}

\item{in_cutoff}{A numeric value less than 1. Signatures from within \code{W}
with an overall exposure less than \code{in_cutoff} will be
discarded for the clustering.}

\item{in_filename}{A path to save the heatmap. If none is specified, the figure will be plotted
to the running environment.}

\item{in_title}{Title in the figure to be created under \code{in_filename}}
}
\value{
The function doesn't return any value.
}
\description{
The PIDs are clustered according to their signature exposures. The procedure
is analogous to \code{\link{hclust_exposures}}, but the graphical output is
more detailed due to the heatmap. This function calls:
\itemize{
 \item \code{\link[gplots]{heatmap.2}} for all statistics.
}
}
\examples{
data(lymphoma_cohort_LCD_results)
 heatmap_exposures(rel_lymphoma_Nature2013_COSMIC_cutoff_exposures_df,
                   COSMIC_subgroups_df,
                   chosen_signatures_indices_df,
                   in_subgroup_colour_column="col",
                   in_method="manhattan",
                   in_subgroup_column="subgroup")
}
\seealso{
\code{\link{hclust}}

\code{\link{dist}}

\code{\link[dendextend]{labels_colors}}
}

