% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCD.R
\name{LCD_cutoff}
\alias{LCD_cutoff}
\title{LCD with a cutoff on exposures}
\usage{
LCD_cutoff(in_mutation_catalogue_df, in_signatures_df, in_cutoff = 0.01,
  in_filename = NULL, in_method = "abs", in_convention = "weak",
  in_per_sample_cutoff = 0)
}
\arguments{
\item{in_mutation_catalogue_df}{A numeric data frame \code{V} with \code{n} rows and \code{m} columns, 
\code{n} being the number of features and \code{m} being the number of samples}

\item{in_signatures_df}{A numeric data frame \code{W} with \code{n} rows and \code{l} columns,
\code{n} being the number of features and \code{l} being the number of signatures}

\item{in_cutoff}{A numeric value less than 1. Signatures from within \code{W}
with an overall exposure less than \code{in_cutoff} will be
discarded.}

\item{in_filename}{A path to generate a histogram of the signature exposures if non-NULL}

\item{in_method}{Indicate to which data the cutoff shall be applied: absolute exposures, relative 
exposures}

\item{in_convention}{Indicate whether LCD or LCD_strict should be used}

\item{in_per_sample_cutoff}{A numeric value less than 1. Signatures from within \code{W}
with an exposure per sample less than \code{in_cutoff} will be
discarded.}
}
\value{
A list with entries \code{exposures}, \code{signatures}, \code{choice}
         and \code{order}
\itemize{
 \item \code{exposures}:
   The exposures \code{H}, a numeric data frame with 
   \code{l} rows and \code{m} columns, \code{l} being
   the number of signatures and \code{m} being the number
   of samples
 \item \code{signatures}:
   The reduced signatures that have exposures bigger
   than \code{in_cutoff}
 \item \code{choice}:
   Index vector of the reduced signatures in the input
   signatures
 \item \code{order}: Order vector of the signatures by exposure
 \item \code{residual_catalogue}:
   Numerical data frame (matrix) of the difference between fit (product of
   signatures and exposures) and input mutational catalogue
 \item \code{rss}:
   Residual sum of squares (i.e. sum of squares of the residual catalogue)
 \item \code{cosDist_fit_orig_per_matrix}:
   Cosine distance between the fit (product of signatures and exposures) and
   input mutational catalogue computed after putting the matrix into vector
   format (i.e. one scaler product for the whole matrix)
 \item \code{cosDist_fit_orig_per_col}:
   Cosine distance between the fit (product of signatures and exposures) and
   input mutational catalogue computed per column (i.e. per sample, i.e. as
   many scaler products as there are samples in the cohort)
 \item \code{sum_ind}:
   Decreasing order of mutational loads based on the input mutational
   catalogue
}
}
\description{
\code{LCD_cutoff} performs a mutational signatures decomposition by 
Linear Combination Decomposition (LCD) of a given
mutational catalogue \code{V} with known signatures \code{W} by 
solving the minimization problem \eqn{min(||W*H - V||)} 
with additional constraints of non-negativity on H where W and V
are known, but excludes signatures with an overall contribution less than
a given cutoff over the whole cohort.
}
\examples{
## define raw data
W_prim <- matrix(c(1,2,3,4,5,6),ncol=2) 
W_prim_df <- as.data.frame(W_prim)
W_df <- YAPSA:::normalize_df_per_dim(W_prim_df,2) # corresponds to the signatures
W <- as.matrix(W_df) 
## 1. Simple case: non-negativity already in raw data
H <- matrix(c(2,5,3,6,1,9,1,2),ncol=4)
H_df <- as.data.frame(H) # corresponds to the exposures
V <- W \%*\% H # matrix multiplication
V_df <- as.data.frame(V) # corresponds to the mutational catalogue
exposures_small_cutoff_list <- YAPSA:::LCD_cutoff(V_df,W_df,in_cutoff = 0.05)
exposures_big_cutoff_list <- YAPSA:::LCD_cutoff(V_df,W_df,in_cutoff = 0.4)

}
\seealso{
\code{\link{LCD}}

\code{\link[limSolve]{lsei}}
}

