% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCD.R
\name{LCD_SMC}
\alias{LCD_SMC}
\title{CD stratification analysis}
\usage{
LCD_SMC(in_mutation_sub_catalogue_list, in_signatures_df, in_F_df = NULL)
}
\arguments{
\item{in_mutation_sub_catalogue_list}{A list of \code{s} stratified
mutational catalogues \code{Vi} \(numeric data frames\) with \code{n} rows
and \code{m} columns each, \code{n} being the number of features and
\code{m} being the number of samples. This list is naturally provided in
\code{\link{run_SMC}}.}

\item{in_signatures_df}{A numeric data frame \code{W} with \code{n} rows and
\code{l} columns, \code{n} being the number of features and \code{l} being
the number of signatures}

\item{in_F_df}{Default NULL}
}
\value{
Returns a list with all exposures and the stratified ones
}
\description{
CD stratification analysis
}
