% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCD.R
\name{LCD_strict}
\alias{LCD_strict}
\title{Linear Combination Decomposition with strict constraints}
\usage{
LCD_strict(in_mutation_catalogue_df, in_signatures_df)
}
\arguments{
\item{in_mutation_catalogue_df}{A numeric data frame \code{V} with \code{n}
rows and \code{m} columns, 
\code{n} being the number of features 
and \code{m} being the number of samples}

\item{in_signatures_df}{A numeric data frame \code{W} with \code{n} rows and 
\code{l} columns, \code{n} being 
the number of features and \code{l} being the
number of signatures}
}
\value{
The exposures \code{H}, a numeric data frame with \code{l} rows and
         \code{m} columns, \code{l} being the number of signatures and
         \code{m} being the number of samples
}
\description{
\code{LCD} performs a mutational signatures decomposition like \code{\link{LCD}}
but with the additional requirement that all residuals be positive.
}
\seealso{
\code{\link{LCD}}

\code{\link{LCD_cutoff}}

\code{\link[limSolve]{lsei}}
}

