% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{lymphoma_Nature2013_raw_df}
\alias{lymphoma_Nature2013_raw_df}
\title{Example data for the vignette}
\usage{
data(lymphoma_Nature2013_raw)
}
\value{
A data frame
}
\description{
A data frame carrying point mutation calls. It represents a subset of the 
data stored in
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/somatic_mutation_data/Lymphoma B-cell/Lymphoma B-cell_clean_somatic_mutations_for_signature_analysis.txt}.
In the file available under that link somatic point mutation calls from 
several samples are listed in a vcf-like format. One column encodes the 
sample the variant was found in.
}
\examples{
data(lymphoma_Nature2013_raw)
head(lymphoma_Nature2013_raw_df)
dim(lymphoma_Nature2013_raw_df)
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23945592}
}

