% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{lymphoma_PID_df}
\alias{lymphoma_PID_df}
\title{Subgroup information for some samples in the vignette}
\usage{
data(lymphoma_PID)
}
\description{
A data frame carrying subgroup information for a subcohort of samples used in the vignette.
Data in the vignette is downloaded from \url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/somatic_mutation_data/Lymphoma B-cell/Lymphoma B-cell_clean_somatic_mutations_for_signature_analysis.txt}.
In the file available under that link somatic point mutation calls from several samples are 
listed in a vcf-like format. One column encodes the sample the variant was found in. In the
vignette we want to restrict the analysis to only a fraction of these involved samples. The
data frame \code{lymphoma_PID_df} carries the sample identifiers (PID) as rownames and the
attributed subgroup in a column called \code{subgroup}.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23945592}
}

