% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{lymphoma_test_df}
\alias{lymphoma_test_df}
\title{Test data for complex functions}
\usage{
data(lymphoma_test)
}
\value{
A data frame
}
\description{
A data frame carrying point mutation calls. It represents a subset of the data stored in
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/somatic_mutation_data/Lymphoma B-cell/Lymphoma B-cell_clean_somatic_mutations_for_signature_analysis.txt}.
In the file available under that link somatic point mutation calls from several samples are 
listed in a vcf-like format. One column encodes the sample the variant was found in. The
data frame \code{lymphoma_test_df} has only the variants occuring in the sample identifiers
(PIDs) 4112512, 4194218 and 4121361.
}
\examples{
data(lymphoma_test)
head(lymphoma_test_df)
dim(lymphoma_test_df)
table(lymphoma_test_df$PID)
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23945592}
}

