% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lymphomaNature2013_mutCat_df}
\alias{lymphomaNature2013_mutCat_df}
\title{Example mutational catalog for the SNV vignette}
\usage{
data(lymphomaNature2013_mutCat_df)
}
\value{
A data fame in the layout of a SNV mutational catalog
}
\description{
\code{lymphomaNature2013_mutCat_df}: A data frame in the format of a SNV mutation catalog. 
The mutational catalog contains SNV variants from the 
\code{lymphoma_Nature2013_raw_df} data. Mutational catalog was created with 
\code{create_mutation_catalogue_from_df} function.
}
\examples{
data(lymphomaNature2013_mutCat_df)
head(lymphomaNature2013_mutCat_df)
dim(lymphomaNature2013_mutCat_df)

}
\references{
paste0("ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/",
"somatic_mutation_data/Lymphoma B-cell/",
"Lymphoma B-cell_clean_somatic_mutations_",
"for_signature_analysis.txt")
}
