% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{melt_exposures}
\alias{melt_exposures}
\title{Generically melts exposure data frames}
\usage{
melt_exposures(in_df)
}
\arguments{
\item{in_df}{Numeric data frame with exposures.}
}
\value{
A data frame with the molten exposures.
}
\description{
Melt an exposure data frame with signatures as ID variables.
}
\examples{
NULL
}

