% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{merge_exposures}
\alias{merge_exposures}
\title{Add an element as first entry to a list}
\usage{
merge_exposures(in_exposures_list, in_signatures_df)
}
\arguments{
\item{in_exposures_list}{List of data frames (carrying information on exposures).}

\item{in_signatures_df}{Data frame \code{W} in which the columns represent the signatures.}
}
\value{
A data frame with the merged exposures.
}
\description{
Merges with the special feature of preserving the signatures and signature
order.
}
\examples{
NULL
}

