% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{normalizeMotifs_otherRownames}
\alias{normalizeMotifs_otherRownames}
\title{Normalize Somatic Motifs with different rownames}
\usage{
normalizeMotifs_otherRownames(in_matrix, in_norms, adjust_counts = TRUE)
}
\arguments{
\item{in_matrix, in_norms}{Arguments to \code{\link[SomaticSignatures]{normalizeMotifs}}}

\item{adjust_counts}{Whether to rescale the counts after adaption or not. Default is true.}
}
\value{
The matrix returned by \code{\link[SomaticSignatures]{normalizeMotifs}}, 
 but with rownames transformed back to the convention of the input
}
\description{
This is a wrapper function to 
\code{\link[SomaticSignatures]{normalizeMotifs}}. The rownames
are first transformed to fit the convention of the 
\code{\link[SomaticSignatures]{SomaticSignatures}}
package and then passed on to the above mentioned function.
}
\examples{
 NULL
 
}

