% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_colour_legend_subgroups}
\alias{plot_colour_legend_subgroups}
\title{Make an annotation plot for subgroup information}
\usage{
plot_colour_legend_subgroups(in_subgroup_df, mycol,
  in_subgroup.field = "subgroup", in_index.field = "index", in_size = 0.9)
}
\arguments{
\item{in_subgroup_df}{Data frame carrying the subgroup information}

\item{mycol}{Vector containing the colour information for the subgroups}

\item{in_subgroup.field}{Indicates the name of the column in which the subgroup information
is encoded}

\item{in_index.field}{Indicates the name of the column in which the index information
is encoded}

\item{in_size}{Numerical value indicating the size unit (important when uniting different
plots)}
}
\value{
The plot as a ggplot2.
}
\description{
For cohort-wide plots like those produced by \code{\link{plot_exposures}}
or \code{\link{plot_relative_exposures}}, annotation plots for per-PID
information can be put on top. This function produces an annotation plot
for subgroup inforamtion.
}
\examples{
 data(lymphoma_cohort_LCD_results)
 temp_anno_color <- aggregate(col~subgroup,data=COSMIC_subgroups_df,
                              function(l) return(l[1]))
 plot_colour_legend_subgroups(COSMIC_subgroups_df,temp_anno_color$col)

}
\seealso{
\code{\link{plot_exposures}}

\code{\link{plot_relative_exposures}}
}

