% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_SMCs.R
\name{plot_strata}
\alias{plot_strata}
\title{Plot all strata from different stratification axes together}
\usage{
plot_strata(in_strata_list, in_signatures_ind_df, output_path = NULL,
  in_attribute = "")
}
\arguments{
\item{in_strata_list}{Data structure created by \code{make_strata_df} or
\code{make_catalogue_strata_df} in which the strata from different
orthogonal stratification axes are reorganized in a consistent
structure.}

\item{in_signatures_ind_df}{A data frame containing meta information about the signatures}

\item{output_path}{Path to directory where the results, especially the figure produced,
are going to be stored.}

\item{in_attribute}{Additional string for the file name where the figure output
is going to be stored.}
}
\value{
The function doesn't return any value.
}
\description{
Plot the cohort wide signature exposures of all strata from 
different stratification axes together. Naturally called by
\code{\link{compare_SMCs}}.
}
\seealso{
\code{\link{compare_SMCs}}
}

