% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{plotExposuresConfidence}
\alias{plotExposuresConfidence}
\title{Plot exposures including confidence intervals}
\usage{
plotExposuresConfidence(in_complete_df, in_subgroups_df, in_sigInd_df)
}
\arguments{
\item{in_complete_df}{Melted numeric input data frame e.g. as computed by
\link{variateExp}}

\item{in_subgroups_df}{Data frame containing meta information on subgroup
attribution of the samples in the cohort of interest.}

\item{in_sigInd_df}{Data frame with meta information on the signatures used
in the analysis.}
}
\value{
The function doesn't return any value but plots instead.
}
\description{
Plot the exposures to extracted signatures including confidence intervals
computed e.g. by \link{variateExp}.
}
\examples{
NULL
}
