% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_functions.R
\name{plotExposuresConfidence_indel}
\alias{plotExposuresConfidence_indel}
\title{Plot exposures including confidence intervals for exposures of SNVs and
INDELs}
\usage{
plotExposuresConfidence_indel(in_complete_df, in_subgroups_df, in_sigInd_df)
}
\arguments{
\item{in_complete_df}{Melted numeric input data frame e.g. as computed by
\code{\link{variateExp}}}

\item{in_subgroups_df}{Data frame containing meta information on subgroup
attribution of the samples in the cohort of interest.}

\item{in_sigInd_df}{Data frame with meta information on the signatures used
in the analysis.}
}
\value{
The function returns a gtable object which can be plotted with
  \code{plot} or \code{grid.draw}
}
\description{
Plot the exposures to extracted signatures including the confidence intervals
computed e.g. by \code{\link[YAPSA]{variateExp}}
}
\examples{
NULL

}
