% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{rel_lymphoma_Nature2013_COSMIC_cutoff_exposures_df}
\alias{rel_lymphoma_Nature2013_COSMIC_cutoff_exposures_df}
\title{Test normalized exposures for plot functions}
\usage{
data(lymphoma_cohort_LCD_results)
}
\value{
A data frame
}
\description{
Data frame with normalized or relative exposures for testing the plot functions.
Data taken from
\url{ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/somatic_mutation_data/Lymphoma B-cell/Lymphoma B-cell_clean_somatic_mutations_for_signature_analysis.txt}.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23945592}
}
\seealso{
\code{\link{lymphoma_Nature2013_COSMIC_cutoff_exposures_df}}

\code{\link{COSMIC_subgroups_df}}

\code{\link{chosen_AlexInitialArtif_sigInd_df}}
}

