% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{relateSigs}
\alias{relateSigs}
\title{Make unique assignments between sets of signatures}
\usage{
relateSigs(querySigs, subjectSigs)
}
\arguments{
\item{querySigs}{The signatures to compare to (given signatures).}

\item{subjectSigs}{The signatures to be compared (new signatures).}
}
\value{
A list of comparison vectors
}
\description{
Make unique assignments between a set of given signatures and a set of new
signatures.
}
\examples{
 NULL

}
\seealso{
\code{\link{compare_sets}}

\code{\link{disambiguateVector}}
}
