% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{repeat_df}
\alias{repeat_df}
\title{Create a data frame with default values}
\usage{
repeat_df(in_value, in_rows, in_cols)
}
\arguments{
\item{in_value}{Default entry to be repeated in the data frame}

\item{in_rows,in_cols}{Dimensions of the data frame to be created}
}
\value{
The created data frame
}
\description{
Create a data frame with default values
}
\examples{
## 1. Initialize with numeric value:
repeat_df(1,2,3)
## 2. Initialize with NA value:
repeat_df(NA,3,2)
## 3. Initialize with character:
repeat_df("a",4,3)
}

