% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_SMCs.R
\name{run_plot_strata_general}
\alias{run_plot_strata_general}
\title{Wrapper function for \code{plot_strata}}
\usage{
run_plot_strata_general(in_stratification_lists_list, in_signatures_ind_df,
  output_path = NULL, in_attribute = "", in_remove_signature_ind = NULL,
  in_additional_stratum = NULL)
}
\arguments{
\item{in_stratification_lists_list}{List of lists with entries from different (orthogonal) stratification
axes or SMCs}

\item{in_signatures_ind_df}{A data frame containing meta information about the signatures}

\item{output_path}{Path to directory where the results, especially the figure produced by
\code{\link{plot_strata}} is going to be stored.}

\item{in_attribute}{Additional string for the file name where the figure produced by
\code{\link{plot_strata}} is going to be stored.}

\item{in_remove_signature_ind}{Omit one of the signatures in \code{in_signatures_ind_df} for the
comparison if non-NULL. The parameter specifies the index of the
signature to be removed.}

\item{in_additional_stratum}{Include an additionally supplied stratum in comparison in non-NULL.}
}
\value{
The function doesn't return any value.
}
\description{
First calls
\itemize{
 \item \code{make_strata_df}, then
 \item \code{\link{plot_strata}}
}
}
\seealso{
\code{\link{plot_strata}}
}

