% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sd_over_present}
\alias{sd_over_present}
\title{Standard deviation of a data frame over a specified dimension}
\usage{
sd_over_present(in_df, in_dimension)
}
\arguments{
\item{in_df}{Data frame the standard deviation is computed of}

\item{in_dimension}{Dimension along which the computation will be carried out}
}
\value{
A vector of the standard deviations
}
\description{
If computing the standard deviation over columns, zero rows
(i.e. those with \code{rowSums=0}) are left out, if computing
the standard deviation over rows, zero columns (i.e. those with
\code{colSums=0}) are left out.
}
\examples{
test_df <- data.frame(matrix(c(1,2,3,0,5,2,3,4,0,6,0,0,0,0,0,4,5,6,0,7),
                             ncol=4))
## 1. Compute standard deviation over non-zero rows:
sd_over_present(test_df,1)
## 2. Compute standard deviation over non-zero columns:
sd_over_present(test_df,2)

}

