% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{stderrmean}
\alias{stderrmean}
\title{Compute the standard error of the mean}
\usage{
stderrmean(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
Standard deviation of an input numerical vector divided by the square
 root of the length of the input vector
}
\description{
This function returns the standard deviation of an input numerical vector
divided by the square root of the length of the input vector
}
\examples{
A <- c(1,2,3)
sd(A)
stderrmean(A)
}

