% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{stratify_and_create_mutational_catalogue}
\alias{stratify_and_create_mutational_catalogue}
\title{Stratify a mutational catalogue and create the data structure}
\usage{
stratify_and_create_mutational_catalogue(our_table, our_column_name,
  vcf_target_dir, strata_target_dir, our_refGenome_Seqinfo,
  our_seqnames.field = "X.CHROM", our_start.field = "POS",
  our_end.field = "POS", our_PID.field = "PID",
  our_subgroup.field = "subgroup", our_refGenome, our_wordLength,
  in_verbose = 1, our_rownames = c())
}
\arguments{
\item{our_table}{A big dataframe \code{T} constructed from a vcf-like file of a whole 
cohort. The first columns are those of a standard vcf file, followed by
an arbitrary number of custom or user defined columns. One of these
can carry a PID (patient or sample identifyier) and one must be the
category used for stratification.}

\item{our_column_name}{Name of the column in \code{in_table} which is going to be used for
stratification}

\item{vcf_target_dir}{Path to the directory where the stratified vcf-like data frames should
be saved if non-NULL}

\item{strata_target_dir}{Path to the directory where the stratified mutational catalogues should
be saved if non-NULL}

\item{our_refGenome_Seqinfo}{A seqInfo object, referring to the reference genome used.
Argument passed on to \code{\link{makeVRangesFromDataFrame}} and thus 
indirectly to \code{\link{makeGRangesFromDataFrame}}.}

\item{our_seqnames.field}{Indicates the name of the column in which the chromosome is encoded}

\item{our_start.field}{Indicates the name of the column in which the start coordinate is
encoded}

\item{our_end.field}{Indicates the name of the column in which the end coordinate is
encoded}

\item{our_PID.field}{Indicates the name of the column in which the PID (patient or sample
identifier) is encoded}

\item{our_subgroup.field}{Indicates the name of the column in which the subgroup information
is encoded}

\item{our_refGenome}{FaFile of the reference genome used.}

\item{our_wordLength}{Integer number defining the length of the features or motifs, e.g. 3 for
tripletts or 5 for pentamers}

\item{in_verbose}{Verbose if \code{in_verbose=1}}

\item{our_rownames}{Optional parameter to specify rownames of the mutational catalogue \code{V}
i.e. the names of the features.}
}
\value{
A list with entries
 \code{table_list},
 \code{name_list} and
 \code{results_lists_list}.
\itemize{
 \item \code{table_list}:
   This is a list of \code{s} dataframes \code{Ti}, where \code{s} is
   the number of strata. The concatenation of all elements of
   \code{table_list} would yield a reordered version of \code{our_table}.
   \code{table_list} is directly taken from the result of
   \code{\link{stratify_vcf_like_df}}
 \item \code{name_list}:
   A list of the names of the different strata equivalent
   to \code{unique(in_table[,in_column_name])}
 \item \code{results_lists_list}
   A list of \code{s} lists, where \code{s is the number of strata}.
   Every entry corresponds to one result of
   \code{\link{create_mutation_catalogue_from_df}}
}
}
\description{
Wrapper function. It first calls
\itemize{
 \item \code{\link{stratify_vcf_like_df}}, then
 \item \code{save_stratified_vcf_like_df}, then
 \item \code{\link{create_mutation_catalogue_from_df}}
   iteratively and finally
 \item \code{save_mutation_catalogue}
}
}
\examples{
NULL
\dontrun{
 library(BSgenome.Hsapiens.UCSC.hg19)
 data(lymphoma_test)
 word_length <- 3
 strata_list <- cut_breaks_as_intervals(
   lymphoma_test_df$random_norm,
   in_outlier_cutoffs=c(-4,4),
   in_cutoff_ranges_list=list(c(-2.5,-1.5),c(0.5,1.5)),
   in_labels=c("small","intermediate","big"))
 lymphoma_test_df$random_cat <- strata_list$category_vector
 temp_list <- stratify_and_create_mutational_catalogue(
   lymphoma_test_df,"random_cat",
   vcf_target_dir=NULL,strata_target_dir=NULL,
   our_refGenome_Seqinfo=NULL,
   our_seqnames.field="CHROM",our_start.field="POS",
   our_end.field="POS",our_PID.field="PID",
   our_subgroup.field="SUBGROUP",
   our_refGenome=BSgenome.Hsapiens.UCSC.hg19,
   our_wordLength=word_length,in_verbose=1,our_rownames=c())
 temp_list$name_list
 names(temp_list$results_lists_list[[1]])
}

}
\seealso{
\code{\link{stratify_vcf_like_df}}

\code{save_stratified_vcf_like_df}

\code{\link{create_mutation_catalogue_from_df}}

\code{save_mutation_catalogue}
}

