% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{stratify_vcf_like_df}
\alias{stratify_vcf_like_df}
\title{Stratify a vcf-like data frame}
\usage{
stratify_vcf_like_df(in_table, in_column_name, in_verbose = 0)
}
\arguments{
\item{in_table}{A big dataframe \code{T} constructed from a vcf-like file of a whole 
cohort. The first columns are those of a standard vcf file, followed by
an arbitrary number of custom or user defined columns. One of these
can carry a PID (patient or sample identifyier) and one must be the
category used for stratification.}

\item{in_column_name}{Name of the column in \code{in_table} which is going to be used for
stratification}

\item{in_verbose}{Verbose if \code{in_verbose=1}}
}
\value{
A list with entries
 \code{table_list} and
 \code{name_list}.
\itemize{
 \item \code{table_list}:
   This is a list of \code{s} dataframes \code{Ti}, where \code{s} is
   the number of strata. The concatenation of all elements of
   \code{table_list} would yield a reordered version of \code{in_table}
 \item \code{name_list}:
   A list of the names of the different strata equivalent
   to \code{unique(in_table[,in_column_name])}
}
}
\description{
Function to cut a big vcf-like data frame \code{T} into a collection of
different strata, i.e. smaller vcf-like data frames \code{Ti}, according
to some attribution in a specified field or column.
}
\examples{
NULL
\dontrun{
 data(lymphoma_test)
 strata_list <- cut_breaks_as_intervals(lymphoma_test_df$random_norm,
                                        in_outlier_cutoffs=c(-4,4),
                                        in_cutoff_ranges_list=list(c(-2.5,-1.5),c(0.5,1.5)),
                                        in_labels=c("small","intermediate","big"))
 lymphoma_test_df$random_cat <- strata_list$category_vector
 stratification_list <- stratify_vcf_like_df(lymphoma_test_df,"random_cat",in_verbose=0)
 stratification_list$name_list
 for (i in seq_len(3)){
   cat("\\nStratum ",i,"\\n")
   cat(dim(stratification_list$table_list[[i]]),"\\n")
 }
}

}
\seealso{
\code{\link{stratify_and_create_mutational_catalogue}}
}

