% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sum_over_list_of_df}
\alias{sum_over_list_of_df}
\title{Elementwise sum over a list of (numerical) data frames}
\usage{
sum_over_list_of_df(in_df_list)
}
\arguments{
\item{in_df_list}{List of (numerical) data frames}
}
\value{
A numerical data frame with the same dimensions as the entries of
 \code{in_df_list} with elementwise sums
}
\description{
Elementwise sum over a list of (numerical) data frames
}
\examples{
A <- data.frame(matrix(c(1,1,1,2,2,2),ncol=2))
B <- data.frame(matrix(c(3,3,3,4,4,4),ncol=2))
df_list <- list(A=A,B=B)
sum_over_list_of_df(df_list)

}

