% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\docType{data}
\name{targetCapture_cor_factors}
\alias{targetCapture_cor_factors}
\title{Correction factors for different target capture kits}
\usage{
data(targetCapture_cor_factors)
}
\description{
List of lists with correction factors for different target capture kits.
The elements of the overall list are lists, every one carrying information
for one target capture kit (and namend after it). The elements of these
sublists are 64 dimensional vectors with correction factors for all
triplets. They were computed using counts of occurence of the respective
triplets in the target capture and in the reference genome and making
ratios (either for the counts themselves as in \code{abs_cor} or for the
relative occurences in \code{rel_cor}). The information in this data
structure may be used as input to
\code{\link{normalizeMotifs_otherRownames}}.
}
\author{
Daniel Huebschmann \email{huebschmann.daniel@googlemail.com}
}

