% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{testSigs}
\alias{testSigs}
\title{Test for significance of alternative models cohort wide}
\usage{
testSigs(
  in_catalogue_df,
  in_sig_df,
  in_exposures_df,
  in_factor = 0,
  in_pdf = NULL
)
}
\arguments{
\item{in_catalogue_df}{Input numerical data frame of the mutational catalog
of the cohort to be analyzed}

\item{in_sig_df}{Numerical data frame of the signatures used for analysis.}

\item{in_exposures_df}{Input numerical data frame of the exposures computed
for the cohort to be analyzed}

\item{in_factor}{Deviation factor of the altered alternative model.}

\item{in_pdf}{Probability distribution function, parameter passed on to
\link{confIntExp} if NULL assumed to be normal distribution.}
}
\value{
Returns a data frame
}
\description{
Wrapper function for \link{variateExpSingle} for application cohort wide.
}
\examples{
NULL

}
