% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{transform_rownames_deconstructSigs_to_YAPSA}
\alias{transform_rownames_deconstructSigs_to_YAPSA}
\title{Change rownames from one naming convention to another}
\usage{
transform_rownames_deconstructSigs_to_YAPSA(in_rownames, wordLength = 3)
}
\arguments{
\item{in_rownames}{Character vector of input rownames}

\item{wordLength}{Size of the considered motif context}
}
\value{
A character vector of the translated rownames.
}
\description{
Rownames or names of the features used differ between the different contexts
a signature analysis is carried out in. The function
\code{transform_rownames_YAPSA_to_deconstructSigs} changes from the 
convention used in the deconstructSigs package to the one used by the YAPSA
pacakge.
}
\examples{
NULL

}
\seealso{
\code{\link{transform_rownames_R_to_MATLAB}}

\code{\link{transform_rownames_MATLAB_to_R}}
}

