% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complex_functions.R
\name{transform_rownames_R_to_MATLAB}
\alias{transform_rownames_R_to_MATLAB}
\title{Change rownames from one naming convention to another}
\usage{
transform_rownames_R_to_MATLAB(in_rownames, wordLength = 3)
}
\arguments{
\item{in_rownames}{Character vector of input rownames}

\item{wordLength}{Size of the considered motif context}
}
\value{
A character vector of the translated rownames.
}
\description{
Rownames or names of the features used differ between the different contexts
a signature analysis is carried out in. The function
\code{transform_rownames_R_to_MATLAB} changes from the convention used in
the YAPSA pacakge to the one used by Alexandrov et al. in the MATLAB
framework.
}
\examples{
NULL
}
\seealso{
\code{\link{transform_rownames_MATLAB_to_R}}

\code{\link{transform_rownames_nature_to_R}}
}

