% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{translate_to_1kG}
\alias{translate_to_1kG}
\title{Translate chromosome names to the 1kG naming convention}
\usage{
translate_to_1kG(in_df, in_CHROM.field = "chr")
}
\arguments{
\item{in_df}{Data frame which carries one column with chromosome information to be 
reformatted}

\item{in_CHROM.field}{String indicating which column of \code{in_df} carries the chromosome 
information}
}
\value{
A data frame identical to \code{in_df}, but with the names in the
 chromosome column replaced
}
\description{
In 1kG, i.e. 1000 genomes naming convention, chromosome names have no prefix
\emph{chr} and the gonosomes are called \emph{23} for \emph{X} and \emph{24}
for \emph{Y}. If data analysis is performed e.g. with \code{hs37d5.fa}, this
naming convention is needed. The inverse transform is done with
\code{\link{translate_to_hg19}}.
}
\examples{
test_df <- data.frame(CHROM=c(1,2,23,24),POS=c(100,120000000,300000,25000),
                      dummy=c("a","b","c","d"))
hg19_df <- translate_to_hg19(test_df, in_CHROM.field = "CHROM")
onekG_df <- translate_to_1kG(hg19_df, in_CHROM.field = "CHROM")
onekG_df

}
\seealso{
\code{\link{translate_to_1kG}}
}

