% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{vplayout}
\alias{vplayout}
\title{Wrapper function for \code{\link[grid]{viewport}}}
\usage{
vplayout(x, y)
}
\arguments{
\item{x, y}{Passed to \code{\link[grid]{viewport}} as \code{layout.pos.row} and
\code{layout.pos.col}}
}
\value{
The result of the \code{\link[grid]{viewport}} function.
}
\description{
Wrapper function for \code{\link[grid]{viewport}}
}
\examples{
 NULL

}

