% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{YAPSA}
\alias{YAPSA}
\alias{YAPSA-package}
\title{Generate R documentation from inline comments.}
\description{
Yet Another Package for mutational Signature analysis
}
\details{
This package provides functions and routines useful in 
the analysis of mutational signatures (cf. L. Alexandrov et al., Nature 
2013). In particular, functions to perform a signature analysis with 
known signatures (\code{\link{LCD}} = linear combination decomposition) and 
a signature analysis on stratified mutational catalogue 
(\code{\link{run_SMC}} = stratify mutational catalogue) are provided.
}

