% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoffs_pcawg}
\alias{cutoffs_pcawg}
\alias{cutoffPCAWG_SBS_WGSWES_artifPid_df}
\alias{cutoffPCAWG_SBS_WGSWES_realPid_df}
\alias{cutoffPCAWG_ID_WGS_Pid_df}
\title{Opt. cutoffs, PCAWG SNV signatures, including artifacts}
\usage{
data(cutoffs_pcawg)
}
\description{
\code{cutoffPCAWG_SBS_WGSWES_artifPid_df}: Optimal cutoffs for 
\code{\link{PCAWG_SP_SBS_sigs_Artif_df}}, i.e. initially published 
signatures,including artifact signatures, trained in a single-sample 
resolution.

\code{cutoffPCAWG_SBS_WGSWES_realPid_df}: Optimal cutoffs for 
\code{\link{PCAWG_SP_SBS_sigs_Real_df}}, i.e. initially published 
signatures, only validated signatures, trained in a single-sample 
resolution.

\code{cutoffPCAWG_ID_WGS_Pid_df}: Optimal cutoffs for 
\code{\link{PCAWG_SP_ID_sigs_df}}, i.e. initially published signatures, 
signatures, trained in a single-sample resolution.
}
\author{
Lea Jopp-Saile \email{huebschmann.daniel@googlemail.com}
}
