% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCD.R
\name{LCD_SMC}
\alias{LCD_SMC}
\title{LCD of stratified mutational catalogues}
\usage{
LCD_SMC(in_mutation_sub_catalogue_list, in_signatures_df, in_F_df = NULL)
}
\arguments{
\item{in_mutation_sub_catalogue_list}{A list of \code{s} numeric data frames \code{Vi}, the mutational catalogues 
per stratum, with \code{n} rows and \code{m} columns each, \code{n} being 
the number of features and \code{m} being the number of samples.}

\item{in_signatures_df}{A numeric data frame \code{W} with \code{n} rows and \code{l} columns,
\code{n} being the number of features and \code{l} being the number
of signatures}

\item{in_F_df}{Optional argument, if specified, \code{H}, i.e. the overall exposures
without stratification, is set to equal \code{in_F_df}. This is equivalent
to forcing the LCD_SMC procedure to use e.g. the exposures of a previously
performed NMF decomposition.}
}
\value{
A list with entries \code{exposures_all_df} and 
         \code{sub_exposures_list}
\itemize{
 \item \code{exposures_all_df}:
   The overall exposures \code{H} without stratification, numeric data frame
   with \code{l} rows and \code{m} columns, \code{l} being the number of 
   signatures and \code{m} being the number of samples, solving the 
   minimization \eqn{min(||W*H - V||)}. If the optional parmeter 
   \code{in_F_df} was specified, \code{exposures_all_df} equals 
   \code{in_F_df}.
 \item \code{sub_exposures_list}:
   The list of \code{s} strata specific exposures Hi, all are numerical data 
   frames with \code{l} rows and \code{m} columns, \code{l} being the number
   of signatures and \code{m} being the number of samples
}
}
\description{
\code{LCD_SMC} performs a mutational signatures decomposition by 
Linear Combination Decomposition (LCD) of a given collection of
stratified mutational catalogues \code{Vi} with known signatures \code{W} by 
simultaneously solving the minimization problems \eqn{min(||W*Hi - Vi||)}
for all \code{i} and \eqn{min(||W*H - V||)} with the following additional 
constraints
\enumerate{
 \item non-negativity of Hi 
 \item \eqn{\Sigma Hi = H}
}
where W and Vi are known.
}
\examples{
## define raw data
W_prim <- matrix(c(1,2,3,4,5,6),ncol=2) 
W_prim_df <- as.data.frame(W_prim)
W_df <- YAPSA:::normalize_df_per_dim(W_prim_df,2) # corresponds to the sigs
W <- as.matrix(W_df) 
H <- matrix(c(2,5,3,6,1,9,1,2),ncol=4)
## define indices where sign is going to be swapped for different strata 
## (perturbation)
sign_ind_1 <- c(5,7)
sign_ind_2 <- c(1)
## alter the matrix H to yield a new mutational catalogue for every stratum 
## (perturbation)
H_1 <- H
H_2 <- H
H_1[sign_ind_1] <- (-1)*H[sign_ind_1]
H_2[sign_ind_2] <- (-1)*H[sign_ind_2]
H_1_df <- as.data.frame(H_1)
H_2_df <- as.data.frame(H_2)
V_1 <- W \%*\% H_1 # matrix multiplication
V_2 <- W \%*\% H_2 # matrix multiplication
V <- V_1 + V_2
V_1_df <- as.data.frame(V_1) # mutational catalog of stratum 1
V_2_df <- as.data.frame(V_2) # mutational catalog of stratum 2
V_df <- as.data.frame(V) # mutational catalog of the whole cohort
V_list <- list()  # make list of data frames
V_list[[1]] <- V_1_df
V_list[[2]] <- V_2_df
## apply function
exposures_strata_list <- YAPSA:::LCD_SMC(V_list,W_df)
## print content of result
exposures_strata_list$exposures_all_df
exposures_strata_list$sub_exposures_list[[1]]
exposures_strata_list$sub_exposures_list[[2]]
## compare
simple_exposures_all_df <- YAPSA:::LCD(V_df,W_df)
simple_exposures_1_df <- YAPSA:::LCD(V_1_df,W_df)
simple_exposures_1 <- as.matrix(simple_exposures_1_df)
simple_exposures_2_df <- YAPSA:::LCD(V_2_df,W_df)
simple_exposures_2 <- as.matrix(simple_exposures_2_df)

}
\seealso{
\code{\link{LCD}}

\code{\link[lsei]{lsei}}
}

