% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{normalize_df_per_dim}
\alias{normalize_df_per_dim}
\title{Normalize a numerical data frame over a specified dimension}
\usage{
normalize_df_per_dim(in_df, in_dimension)
}
\arguments{
\item{in_df}{Data frame to be normalized}

\item{in_dimension}{Dimension along which the normalization will be carried out}
}
\value{
The normalized numerical data frame
}
\description{
Normalization is carried out by dividing by \code{rowSums} or \code{colSums};
for rows with \code{rowSums=0} or columns with \code{colSums=0}, the
normalization is left out.
}
\examples{
test_df <- data.frame(matrix(c(1,2,3,0,5,2,3,4,0,6,0,0,0,0,0,4,5,6,0,7),
                             ncol=4))
## 1. Normalize over rows:
normalize_df_per_dim(test_df,1)
## 2. Normalize over columns:
normalize_df_per_dim(test_df,2)

}

